se crea componente createUser para crear usuarios
